Organize file difference utility functions
