+++ .changeset/chilled-ties-speak.md
@@ -1,0 +1,5 @@
+---
+"@osrs-wiki/cache-mediawiki": patch
+---
+
+Organize file difference utility functions
