+++ src/scripts/differences/file/content/enums.ts
@@ -2,7 +2,7 @@ import _ from "underscore";
 
 import { Enum, EnumID, Reader } from "../../../../utils/cache2";
 import { CompareFn } from "../../differences.types";
-import { getFileDifferences } from "../file";
+import { getFileDifferences } from "../file.utils";
 
 const compareEnums: CompareFn = ({ oldFile, newFile }) => {
   const oldEntry = oldFile
