+++ src/scripts/differences/file/content/npcs.ts
@@ -4,7 +4,7 @@ import Context from "../../../../context";
 import { NPC, NPCID, Reader } from "../../../../utils/cache2";
 import { buildNpcInfobox } from "../../../infoboxGenernator/infoboxes/npc";
 import { CompareFn } from "../../differences.types";
-import { getFileDifferences } from "../file";
+import { getFileDifferences } from "../file.utils";
 
 const compareNpcs: CompareFn = ({ oldFile, newFile }) => {
   const oldEntry = oldFile
