+++ src/scripts/differences/file/content/objects.ts
@@ -2,7 +2,7 @@ import _ from "underscore";
 
 import { Obj, ObjID, Reader } from "../../../../utils/cache2";
 import { CompareFn } from "../../differences.types";
-import { getFileDifferences } from "../file";
+import { getFileDifferences } from "../file.utils";
 
 const compareObjects: CompareFn = ({ oldFile, newFile }) => {
   const oldEntry = oldFile
