+++ src/scripts/differences/file/content/params.ts
@@ -2,7 +2,7 @@ import _ from "underscore";
 
 import { Param, ParamID, Reader } from "../../../../utils/cache2";
 import { CompareFn } from "../../differences.types";
-import { getFileDifferences } from "../file";
+import { getFileDifferences } from "../file.utils";
 
 const compareParams: CompareFn = ({ oldFile, newFile }) => {
   const oldEntry = oldFile
